requirements.txt
aiofiles==23.1.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.11.7
asttokens==2.2.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
black==22.6.0
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
dill==0.3.5.1
Django==4.1
dnspython==2.2.1
email-validator==1.2.1
executing==1.2.0
flake8==5.0.4
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Compress==1.12
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.34.4
frozenlist==1.3.1
getsize==1.1.11
greenlet==1.1.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==8.8.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lazy_loader==0.1rc2
llvmlite==0.39.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.2
multipledispatch==0.6.0
mypy-extensions==0.4.3
natsort==8.1.0
numba==0.56.4
numpy==1.23.1
opencv-python==4.7.0.72
packaging==21.3
pandas==1.4.3
pandas-flavor==0.3.0
pandas-stubs==1.4.3.220815
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pycsp==0.9.2
pydantic==1.9.1
pyflakes==2.5.0
pygame==2.2.0
Pygments==2.14.0
pyjanitor==0.23.1
pylint==2.14.5
Pympler==1.0.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-i18n==0.3.9
pytz==2022.1
regex==2022.9.13
requests==2.28.1
Rx==3.2.0
scipy==1.9.0
six==1.16.0
snakeviz==2.1.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlalchemy-orm==1.2.3
sqlparse==0.4.2
stack-data==0.6.2
tenacity==8.0.1
tk==0.1.0
tomli==2.0.1
tomlkit==0.11.4
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
types-pytz==2022.2.1.0
typing-inspect==0.7.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
xarray==2022.6.0
yarl==1.8.1
